refactor: modularize task list
